# apps/engine/src/__generated__/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stripe Sync Engine",
    "version": "1.0.0",
    "description": "Stripe Sync Engine — stateless, one-shot source/destination sync over HTTP.\nAll sync endpoints accept configuration via the `X-Pipeline` header (JSON-encoded PipelineConfig). Optional cursor state can be provided via `X-State`.\n\n## Endpoints\n\n| Method | Path | Summary |\n|--------|------|---------|\n| GET | /health | Health check |\n| POST | /pipeline_check | Check connector connection |\n| POST | /pipeline_setup | Set up destination schema |\n| POST | /pipeline_teardown | Tear down destination schema |\n| POST | /source_discover | Discover available streams |\n| POST | /pipeline_read | Read records from source |\n| POST | /pipeline_write | Write records to destination |\n| POST | /pipeline_sync | Run sync pipeline (read → write) |\n| GET | /meta/sources | List available source connectors |\n| GET | /meta/sources/{type} | Get source connector spec |\n| GET | /meta/destinations | List available destination connectors |\n| GET | /meta/destinations/{type} | Get destination connector spec |"
  },
  "paths": {
...
    "/pipeline_check": {
      "post": {
        "operationId": "pipeline_check",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Check connector connection",
        "description": "Validates the source/destination config and tests connectivity. Streams NDJSON messages (connection_status, log, trace) tagged with _emitted_by.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "[object Object]": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NDJSON stream of check messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/CheckOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/pipeline_setup": {
      "post": {
        "operationId": "pipeline_setup",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Set up destination schema",
        "description": "Creates destination tables and applies migrations. Streams NDJSON messages (control, log, trace) tagged with _emitted_by.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "[object Object]": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NDJSON stream of setup messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/pipeline_teardown": {
      "post": {
        "operationId": "pipeline_teardown",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Tear down destination schema",
        "description": "Drops destination tables. Streams NDJSON messages (log, trace) tagged with _emitted_by.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "[object Object]": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NDJSON stream of teardown messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/TeardownOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/source_discover": {
      "post": {
        "operationId": "source_discover",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Discover available streams",
        "description": "Streams NDJSON messages (catalog, logs, traces) for the configured source.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "[object Object]": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NDJSON stream of discover messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/pipeline_read": {
      "post": {
        "operationId": "pipeline_read",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Read records from source",
        "description": "Streams NDJSON messages (records, state, catalog). Optional NDJSON body provides live events as input.",
        "parameters": [
          {
            "in": "query",
            "name": "state_limit",
            "schema": {
              "description": "Stop streaming after N state messages.",
              "example": "100",
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 9007199254740991
            },
            "description": "Stop streaming after N state messages."
          },
          {
            "in": "query",
            "name": "time_limit",
            "schema": {
              "description": "Stop streaming after N seconds.",
              "example": "10",
              "type": "number",
              "exclusiveMinimum": 0
            },
            "description": "Stop streaming after N seconds."
          },
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "[object Object]": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          },
          {
            "in": "header",
            "name": "x-state",
            "required": false,
            "description": "JSON-encoded per-stream cursor state",
            "content": {
              "[object Object]": {
                "schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              }
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/x-ndjson": {
              "schema": {
                "$ref": "#/components/schemas/SourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NDJSON stream of sync messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/pipeline_write": {
      "post": {
        "operationId": "pipeline_write",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Write records to destination",
        "description": "Reads NDJSON messages from the request body and writes them to the destination. Pipe /read output as input.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "[object Object]": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-ndjson": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NDJSON stream of write result messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/pipeline_sync": {
      "post": {
        "operationId": "pipeline_sync",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Run sync pipeline (read → write)",
        "description": "Without a request body, reads from the source connector and writes to the destination (backfill mode). With an NDJSON request body, uses the provided messages as input instead of reading from the source (push mode — e.g. piped webhook events).",
        "parameters": [
          {
            "in": "query",
            "name": "state_limit",
            "schema": {
              "description": "Stop streaming after N state messages.",
              "example": "100",
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 9007199254740991
            },
            "description": "Stop streaming after N state messages."
          },
          {
            "in": "query",
            "name": "time_limit",
            "schema": {
              "description": "Stop streaming after N seconds.",
              "example": "10",
              "type": "number",
              "exclusiveMinimum": 0
            },
            "description": "Stop streaming after N seconds."
          },
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "[object Object]": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          },
          {
            "in": "header",
            "name": "x-state",
            "required": false,
            "description": "JSON-encoded per-stream cursor state",
            "content": {
              "[object Object]": {
                "schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              }
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/x-ndjson": {
              "schema": {
                "$ref": "#/components/schemas/SourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NDJSON stream of sync messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/SyncOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckOutput": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ConnectionStatusMessage"
          },
          {
            "$ref": "#/components/schemas/LogMessage"
          },
          {
            "$ref": "#/components/schemas/TraceMessage"
          }
        ],
        "type": "object",
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "connection_status": "#/components/schemas/ConnectionStatusMessage",
            "log": "#/components/schemas/LogMessage",
            "trace": "#/components/schemas/TraceMessage"
          }
        }
      },
      "SetupOutput": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ControlMessage"
          },
          {
            "$ref": "#/components/schemas/LogMessage"
          },
          {
            "$ref": "#/components/schemas/TraceMessage"
          }
        ],
        "type": "object",
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "control": "#/components/schemas/ControlMessage",
            "log": "#/components/schemas/LogMessage",
            "trace": "#/components/schemas/TraceMessage"
          }
        }
      },
...
    }
  }
}